requirements.txt
# Web Scraping
beautifulsoup4==4.12.2
requests==2.31.0
lxml==4.9.3
selenium==4.15.0  # Optional, if JS rendering needed

# Data Processing
pandas==2.1.3
numpy==1.24.3

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2
torch==2.1.0  # For sentence-transformers

# LLM
google-generativeai==0.3.2

# API Server
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Utilities
python-dotenv==1.0.0
tqdm==4.66.1
urllib3==2.1.0

